+++ integrations/github/github/webhook/webhook_processors/deployment_webhook_processor.py
@@ -1,0 +1,72 @@
+from loguru import logger
+from github.helpers.utils import ObjectKind
+from github.clients.client_factory import create_github_client
+
+from github.core.options import SingleDeploymentOptions, SingleEnvironmentOptions
+from port_ocean.core.handlers.port_app_config.models import ResourceConfig
+from port_ocean.core.handlers.webhook.webhook_event import (
+    EventPayload,
+    WebhookEvent,
+    WebhookEventRawResults,
+)
+
+from github.core.exporters.deployment_exporter import RestDeploymentExporter
+from github.core.exporters.environment_exporter import RestEnvironmentExporter
+from github.webhook.webhook_processors.base_repository_webhook_processor import (
+    BaseRepositoryWebhookProcessor,
+)
+
+
+class DeploymentWebhookProcessor(BaseRepositoryWebhookProcessor):
+
+    async def _validate_payload(self, payload: EventPayload) -> bool:
+        return bool(
+            payload.get("deployment", {}).get("id")
+            and payload.get("deployment", {}).get("environment")
+        )
+
+    async def _should_process_event(self, event: WebhookEvent) -> bool:
+        return event.headers.get("x-github-event") in [
+            "deployment",
+            "deployment_status",
+        ]
+
+    async def get_matching_kinds(self, event: WebhookEvent) -> list[str]:
+        return [ObjectKind.DEPLOYMENT, ObjectKind.ENVIRONMENT]
+
+    async def handle_event(
+        self, payload: EventPayload, resource_config: ResourceConfig
+    ) -> WebhookEventRawResults:
+        action = payload["action"]
+        deployment = payload["deployment"]
+        environment = deployment["environment"]
+        deployment_id = str(deployment["id"])
+        repo = payload["repository"]["name"]
+        resource_config_kind = resource_config.kind
+
+        logger.info(
+            f"Processing deployment event: {action} for {resource_config_kind} in {repo}"
+        )
+
+        client = create_github_client()
+
+        if resource_config_kind == ObjectKind.ENVIRONMENT:
+            environment_exporter = RestEnvironmentExporter(client)
+            data_to_upsert = await environment_exporter.get_resource(
+                SingleEnvironmentOptions(
+                    repo_name=repo,
+                    name=environment,
+                )
+            )
+        else:  # ObjectKind.DEPLOYMENT
+            deployment_exporter = RestDeploymentExporter(client)
+            data_to_upsert = await deployment_exporter.get_resource(
+                SingleDeploymentOptions(
+                    repo_name=repo,
+                    id=deployment_id,
+                )
+            )
+
+        return WebhookEventRawResults(
+            updated_raw_results=[data_to_upsert], deleted_raw_results=[]
+        )